Remove proxy models from the list of reportable models (#4)
Proxy models are not "real" tables, they are essentially views into an existing one.

So, like abstract models, they should not be included in the snapshot generation.

This fix ensures that is the case. A proxy model is added to the test model suite to ensure that it does not show up in the existing test suite.